Move some files around for better grouping
