+++ packages/core/src/i18n/i18n-setup.js
