+++ packages/core/src/i18n/language-detector.js
