+++ packages/core/src/i18n/translations/es.js
